% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcHatMatrix.R
\name{CalcHatMatrix}
\alias{CalcHatMatrix}
\title{CalcHatMatrix}
\usage{
CalcHatMatrix(X, true.beta)
}
\arguments{
\item{X}{design matrix}

\item{true.beta}{vector of true betas}
}
\value{
the hat matrix of a design matrix
}
\description{
This function returns the hat matrix according to a design matrix
}
\examples{
CalcHatMatrix(matrix(data=c(1,2,3,2),nrow=2),true.beta=c(1,2))
}
\keyword{fisherinfo}
