% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcLRneg.R
\name{CalcLRneg}
\alias{CalcLRneg}
\title{CalcLRneg}
\usage{
CalcLRneg(zsens, zspec)
}
\arguments{
\item{zsens}{numeric; sensitivity of the binary test}

\item{zspec}{numeric; specificity of the binary test}
}
\value{
the negative likelihood ratio of a binary tests
}
\description{
This function returns the negative likelihood ratio of a binary test
}
\examples{
CalcLRneg(0.2,0.8)

}
\keyword{lrpos}
