% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcLRpos.R
\name{CalcLRpos}
\alias{CalcLRpos}
\title{CalcLRpos}
\usage{
CalcLRpos(zsens, zspec)
}
\arguments{
\item{zsens}{numeric; sensitivity of the binary test}

\item{zspec}{numeric; specificity of the binary test}
}
\value{
the positive likelihood ratio of a binary tests
}
\description{
This function returns the positive likelihood ratio of a binary test
}
\examples{
CalcLRpos(0.2,0.8)

}
\keyword{lrpos}
