% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcNPV.R
\name{CalcNPV}
\alias{CalcNPV}
\title{CalcNPV}
\usage{
CalcNPV(sens, spec, prev)
}
\arguments{
\item{sens}{numeric; sensitivity of the binary test}

\item{spec}{numeric; specificity of the binary test}

\item{prev}{numeric; prevalence of the finding}
}
\value{
the negative predicted value of a binary test
}
\description{
This function returns the negative predicted value of a binary test
}
\examples{
CalcNPV(0.2,0.8,0.13)
}
\keyword{npv}
