% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcOneAucVal.R
\name{CalcOneAucVal}
\alias{CalcOneAucVal}
\title{CalcOneAucVal}
\usage{
CalcOneAucVal(test.df, n_val, stratify_val, nz1, nz0, nlp.model)
}
\arguments{
\item{test.df}{Initial test df}

\item{n_val}{Number of validation samples}

\item{nz1}{Number of z=1 in the weighted bootstrap sample}

\item{nz0}{Number of z=0 in the weighted bootstrap sample}

\item{nlp.model}{The fitted model}

\item{stratify_var}{Surrogate name in quo(.)}
}
\value{
Estimated value of validation AUC based on a single
weighted bootstrap sample
}
\description{
Runs a K-fold cross-validation with weighted bootstrap
to approximate training/testing error from an annotated dataset
}
\keyword{bootstrap}
\keyword{cross}
\keyword{validation,}
\keyword{weighted}
