% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcORatio.R
\name{CalcORatio}
\alias{CalcORatio}
\title{CalcORatio}
\usage{
CalcORatio(zsens, zspec, R = 0.5, pz)
}
\arguments{
\item{zsens}{numeric; sensitivity of the surrogate Z for the outcome}

\item{zspec}{numeric; specificity of the surrogate Z for the outcome}

\item{R}{numeric; ratio of surrogate positives in sample i.e. P(Z=1|S=1)}

\item{pz}{numeric; prevalence of surrogate in cohort i.e. P(Z=1)}
}
\value{
the Oratio of the sample
}
\description{
This function returns the o_ratio a surrogate guided sampling design
for a binary outcome compared to a random sample
}
\examples{
CalcORatio(0.2,0.8,0.5,0.12)

}
\keyword{oratio}
