% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPz.R
\name{CalcPz}
\alias{CalcPz}
\title{CalcPz}
\usage{
CalcPz(sens, spec, prev)
}
\arguments{
\item{sens}{numeric; sensitivity of the binary test}

\item{spec}{numeric; specificity of the binary test}

\item{prev}{numeric; prevalence of the finding}
}
\value{
the marginal probability of a binary test
}
\description{
This function returns the marginal probability of a binary test.
}
\examples{
CalcPz(0.2,0.8,0.13)

}
\keyword{pz}
