% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcTrapezoidalAUC.R
\name{CalcTrapezoidalAUC}
\alias{CalcTrapezoidalAUC}
\title{CalcTrapezoidalAUC}
\usage{
CalcTrapezoidalAUC(sens, spec)
}
\arguments{
\item{sens}{numeric; sensitivity of the binary test}

\item{spec}{numeric; specificity of the binary test}
}
\value{
the area under the ROC curve of the binary test for the binary outcome
}
\description{
This function returns the auc of a binary test 
according to the trapezoidal rule.
}
\examples{
CalcTrapezoidalAUC(0.2,0.8)

}
\keyword{auc}
