% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVWeightedBootstrap.R
\name{CVWeightedBootstrap}
\alias{CVWeightedBootstrap}
\title{CVWeightedBootstrap}
\usage{
CVWeightedBootstrap(feature.mat, k_fold = 5, boot_samples = 100, pz,
  stratify_var = quo(Z), Y = "Y", model = "lasso")
}
\arguments{
\item{feature.mat}{Data frame containing features and labels}

\item{k_fold}{Number of folds for cross validation}

\item{boot_samples}{Number of weighted bootstrap samples for validation}

\item{pz}{Prevalence of surrogate positives in the cohort i.e. P(Z=1)}

\item{stratify_var}{Surrogate name in quo(.)}

\item{Y}{Column name of the binary outcome}

\item{model}{Whether to use a glm, lasso, or ridge model}
}
\value{
A data frame with k_fold rows, where each row has the
AUC's obtained from training, validation (naive, no reweighting),
as well as the mean and variance of the weighted bootstrap resampling
}
\description{
Runs a K-fold cross-validation with weighted bootstrap
to approximate training/testing error from an annotated dataset
}
\keyword{bootstrap}
\keyword{cross}
\keyword{validation,}
\keyword{weighted}
