% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expit.R
\name{Expit}
\alias{Expit}
\title{Expit}
\usage{
Expit(x)
}
\arguments{
\item{x}{numeric; number to take the expit of}
}
\value{
the expit of x
}
\description{
This function returns the expit of a number
}
\keyword{expit}
