% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenCohortMultiOutcome.R
\name{GenCohortMultiOutcome}
\alias{GenCohortMultiOutcome}
\title{GenCohortMultiOutcome}
\usage{
GenCohortMultiOutcome(n, K, p, p_nonzero, beta0, betaZ.mat, betaX, probZ, probX,
  corr_y = NULL)
}
\arguments{
\item{n}{cohort size}

\item{K}{number of multivariate Bernoulli outcomes}

\item{p}{Number of covariates (besides surrogates)}

\item{p_nonzero}{Number of nonzero covariates (besides surrogates)}

\item{beta0}{Length K vectors of intercepts; defines f(Y)}

\item{betaZ.mat}{KxK matrix of coefficients of surrogate; defines f(Z,Y)}

\item{betaX}{Length p vector of coefficient; defines f(X,Y)}

\item{probZ}{Length K vector of probabilities; defines f(Z)}

\item{probX}{Single number; defines f(X)}

\item{corr_y}{Correlation between outcomes, defaults to NULL}
}
\value{
A list with the following components:
}
\description{
Generates a cohort with multivariate bernoulli outcomes
where each outcome has an associated "surrogate"
}
\keyword{data}
\keyword{generate}
