% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logit.R
\name{Logit}
\alias{Logit}
\title{Logit}
\usage{
Logit(y)
}
\arguments{
\item{y}{numeric; number to take the logit of}
}
\value{
the logit of y
}
\description{
This function returns the logit of a number
}
\examples{
Logit()
}
\keyword{logit}
