% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrogateCohortSample.R
\name{SurrogateCohortSample}
\alias{SurrogateCohortSample}
\title{SurrogateCohortSample}
\usage{
SurrogateCohortSample(df, stratify.var = c(quo(Z1), quo(Z2)),
  n.strata = c(400, 400, 0), REPLACE = FALSE)
}
\arguments{
\item{df}{Input data frame}

\item{stratify.var}{Vector of stratification variables in quo()}

\item{n.strata}{Vector of sample sizes for each stratify.var element;
last element is sample size for simple random sample}

\item{REPLACE}{whether to sample with replacement}
}
\value{
sampled.df: A stratified sample with
sum(n.strata) rows according to specified sampling scheme
}
\description{
Creates a stratified sample based on values of stratify_var
Where each partition/stratification has size denoted by n.strata
}
\keyword{cohort}
\keyword{enriched}
\keyword{sample}
\keyword{sample,}
\keyword{surrogate}
