% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrogateGuidedSample.R
\name{SurrogateGuidedSample}
\alias{SurrogateGuidedSample}
\title{SurrogateGuidedSample}
\usage{
SurrogateGuidedSample(df, n_sample = 100, stratify_var = NULL, nz1 = NULL,
  nz0 = NULL, REPLACE = FALSE, IS_SRS = FALSE)
}
\arguments{
\item{df}{Input data frame}

\item{n_sample}{Sample size}

\item{stratify_var}{quo(column in df) to stratify sample on;
defaults to NULL for simple random sample}

\item{nz1}{Number of surrogate positives in sample}

\item{nz0}{Number of surrogate negatives in sample}

\item{REPLACE:}{whether to sample with replacement}
}
\value{
sampled.df: A data frame with length n_sample according to
specified sampling scheme
}
\description{
Makes a sample enriched for cases using ideas from
surrogate guided sampling strategy
}
\keyword{enriched}
\keyword{guided}
\keyword{sample}
\keyword{sample,}
\keyword{surrogate}
