% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{ValidateModel}
\alias{ValidateModel}
\title{ValidateModel}
\usage{
ValidateModel(train.model, val.df, feat = "X", surr = "Z", Y = "Y")
}
\arguments{
\item{train.model}{The trained model object}

\item{val.df}{Data frame containing training data}

\item{Y}{Column name of the binary outcome}
}
\value{
A list of
  eta: vector of linear predictions
  auc_val: Estimated validation AUC
}
\description{
Makes a sample enriched for cases using ideas from
surrogate guided sampling strategy
}
\keyword{model}
\keyword{validate}
