% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenCohortSingleOutcome.R
\name{GenCohortSingleOutcome}
\alias{GenCohortSingleOutcome}
\title{GenCohortSingleOutcome}
\usage{
GenCohortSingleOutcome(n = 1e+05, p = 250, p_nonzero = 30,
  beta0 = -14.75, betaZ = c(4.25, 2.25, 3.35), betaX, probZ = c(0.065,
  0.34, 0.62), probX)
}
\arguments{
\item{n}{cohort size}

\item{p}{number of features}

\item{p_nonzero}{number of nonzero features}

\item{beta0}{intercept; controls prevalence of finding}

\item{betaZ}{vector of true coefficients for surrogate Z}

\item{betaX}{vector of true coefficients for features X}

\item{probZ}{vector of probabilities of potential surrogate, i.e. P(Z=1)}

\item{probX}{vector of probabilities of features, i.e. P(X=1)}
}
\value{
A list with the following components:
X: Matrix of dimension n x p for features
Z: Matrix of dimension n x length(betaZ) for surrogates
Y: Vector of 0/1 labels of length n
trueAUC: AUC of the population if true probabilities were known
true.beta: Vector of true coefficients
}
\description{
Generates a cohort with a single Bernoulli outcome
with p features and length(betaZ) potential surrogates
}
\keyword{data}
\keyword{generate}
