% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainModel.R
\name{TrainModel}
\alias{TrainModel}
\title{TrainModel}
\usage{
TrainModel(train.df, Y = "Y")
}
\arguments{
\item{train.df}{Data frame containing training data}

\item{Y}{Column name of the binary outcome}
}
\value{
A list of
  coef: estimated betas
  n_cases: Number of cases in the training sample
  auc_train: Estimated training AUC
  train.model: The trained model object
}
\description{
Makes a sample enriched for cases using ideas from
surrogate guided sampling strategy
}
\keyword{model}
\keyword{train}
